requirements.txt
torch>=1.3
torchvision>=0.8.2
numpy>=1.16
pillow